+++ rocketgram/keyboards/reply.py
@@ -3,34 +3,43 @@
 # Rocketgram is released under the MIT License (see LICENSE).
 
 
+from typing import Optional
+
 from .keyboard import Keyboard
 from .. import api
 
 
 class ReplyKeyboard(Keyboard):
-    __slots__ = ('__selective', '__one_time', '__resize')
+    __slots__ = ('__resize', '__one_time', '__placeholder', '__selective')
 
-    def __init__(self, *, selective=False, one_time=False, resize=True):
+    def __init__(self, *, resize: bool = True, placeholder: Optional[str] = None, one_time: bool = False,
+                 selective: bool = False):
         super().__init__()
 
-        self.__selective = selective
+        self.__resize = resize
         self.__one_time = one_time
+        self.__placeholder = placeholder
+        self.__selective = selective
+
+    def set_resize(self, resize=False):
         self.__resize = resize
 
-    def set_selective(self, selective=False):
-        self.__selective = selective
-
-    selective = property(fget=lambda self: self.__selective, fset=set_selective)
+    resize = property(fget=lambda self: self.__resize, fset=set_resize)
 
     def set_one_time(self, one_time=False):
         self.__one_time = one_time
 
     one_time = property(fget=lambda self: self.__one_time, fset=set_one_time)
 
-    def set_resize(self, resize=False):
-        self.__resize = resize
+    def set_placeholder(self, placeholder: Optional[str] = None):
+        self.__placeholder = placeholder
 
-    resize = property(fget=lambda self: self.__resize, fset=set_resize)
+    placeholder = property(fget=lambda self: self.__placeholder, fset=set_placeholder)
+
+    def set_selective(self, selective=False):
+        self.__selective = selective
+
+    selective = property(fget=lambda self: self.__selective, fset=set_selective)
 
     def text(self, text: str) -> 'ReplyKeyboard':
         self.add(api.KeyboardButton(text=text))
@@ -55,4 +64,5 @@ class ReplyKeyboard(Keyboard):
         return api.ReplyKeyboardMarkup(self.render_buttons(),
                                        resize_keyboard=self.resize if self.resize else None,
                                        one_time_keyboard=self.one_time if self.one_time else None,
+                                       input_field_placeholder=self.placeholder,
                                        selective=self.selective if self.selective else None)